norm

+++ srcs/caster/calculation.c
@@ -6,7 +6,7 @@
 /*   By: abdeel-o <abdeel-o@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/26 12:35:35 by abdeel-o          #+#    #+#             */
-/*   Updated: 2023/08/27 12:00:35 by abdeel-o         ###   ########.fr       */
+/*   Updated: 2023/08/28 10:44:22 by abdeel-o         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,10 +34,10 @@ void	calc_distance(t_all *all, t_player *p)
 		all->horz_hit_distance = distance_bet_points(p->x, p->y, \
 			all->horz_hit.wall_hit_x, all->horz_hit.wall_hit_y);
 	else
-		all->horz_hit_distance = FLT_MAX;
+		all->horz_hit_distance = INT_MAX;
 	if (all->ver_hit.found_wall_hit)
 		all->ver_hit_distance = distance_bet_points(p->x, p->y, \
 			all->ver_hit.wall_hit_x, all->ver_hit.wall_hit_y);
 	else
-		all->ver_hit_distance = FLT_MAX;
+		all->ver_hit_distance = INT_MAX;
 }
